Change the namespace of the Updater classes
